+++ jest.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   preset: "ts-jest",
+  setupFilesAfterEnv: ["./src/test/custom-matchers.ts"],
   testEnvironment: "node",
   testPathIgnorePatterns: ["/node_modules/", "/out/", ".contract.test.ts"],
   globals: {